move housplant to API, add data filtering
